Adjust to right exception for chunkeddriver
Fix the wrong exception for getting the list of object
names during backup deleting. Since the chunkeddriver has
been refactored out as the top abstract backup driver for
swift and nfs, and there might be other backup backend
coming, it should be fit to the proper backend instead
of hardcoded swift. And update log messages beginning
with capital letter in this driver.

Change-Id: Ia4f9fea11b0fc3c15fc1543663826c5915e26ce7
Signed-off-by: Lei Li <li.lei2@hp.com>